+++ srcs/execution/exec_pipe_list.c
@@ -6,7 +6,7 @@
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/01 21:29:47 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/18 18:56:46 by jnakahod         ###   ########.fr       */
+/*   Updated: 2021/10/18 20:40:53y jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,24 +25,21 @@ static void	child_exec_cmd(t_pipe_list *pipe_list)
 	builtin_id = get_builtin_id(pipe_list->cmd_args[0]);
 	if (builtin_id != NON_BUILTIN)
 		g_request.builtin_funcs[builtin_id](pipe_list->cmd_args, FALSE);
-	environs = env_list_to_array(g_request.environs);
-	cmd_args = pipe_list->cmd_args;
-	if (is_path_part((char *)cmd_args[0]))
+	else
 	{
-		if (!search_path((char **)cmd_args))
-			print_err_and_exit(ERR_MSG_INVLD_CMD, CMD_NOT_FND);
+		environs = env_list_to_array(g_request.environs);
+		cmd_args = pipe_list->cmd_args;
+		if (is_path_part((char *)cmd_args[0]))
+		{
+			if (!search_path((char **)cmd_args))
+				print_err_and_exit(ERR_MSG_INVLD_CMD, CMD_NOT_FND);
+		}
+		if (access(cmd_args[0], X_OK) == -1)
+			print_err_and_exit(NULL, DENIED);
+		if (execve(cmd_args[0], (char *const *)cmd_args, environs) < 0)
+			print_err_and_exit(NULL, GNRL_ERR);
 	}
-	if (access(cmd_args[0], X_OK) == -1)
-		print_err_and_exit(NULL, DENIED);
-	if (execve(cmd_args[0], (char *const *)cmd_args, environs) < 0)
-		print_err_and_exit(NULL, GNRL_ERR);
-}
-
-t_bool	has_pipe(t_pipe_list *pipe_list)
-{
-	if (pipe_list->next)
-		return (TRUE);
-	return (FALSE);
+	exit(0);
 }
 
 void	exec_simple_cmd(t_pipe_list *pipe_list)
@@ -68,60 +65,6 @@ void	exec_simple_cmd(t_pipe_list *pipe_list)
 		g_request.exit_cd = WEXITSTATUS(status);
 		if (changed_pid == -1)
 			perror("waitpid");
-	}
-}
-
-void	init_pipe_fd(int pipe_fd[2])
-{
-	pipe_fd[0] = -1;
-	pipe_fd[1] = -1;
-}
-
-t_bool	is_pipe_open(int pipe_fd[2])
-{
-	if (pipe_fd[0] == -1 && pipe_fd[1] == -1)
-		return (FALSE);
-	return (TRUE);
-}
-
-void	read_pipe(int pipe_fd[2])
-{
-	close(pipe_fd[1]);
-	dup2(pipe_fd[0], 0);
-	close(pipe_fd[0]);
-}
-
-void	write_pipe(int pipe_fd[2])
-{
-	close(pipe_fd[0]);
-	dup2(pipe_fd[1], 1);
-	close(pipe_fd[1]);
-}
-
-void	child_operate_pipe_fd(t_pipe_list *first, t_pipe_list *node, int last_pipe_fd[2], int new_pipe_fd[2])
-{
-	if (!has_pipe(node))
-		read_pipe(last_pipe_fd);
-	else if (first == node)
-		write_pipe(new_pipe_fd);
-	else
-	{
-		read_pipe(last_pipe_fd);
-		write_pipe(new_pipe_fd);
-	}
-}
-
-void	parent_operate_pipe_fd(t_pipe_list *node, int last_pipe_fd[2], int new_pipe_fd[2])
-{
-	if (is_pipe_open(last_pipe_fd))
-	{
-		close(last_pipe_fd[0]);
-		close(last_pipe_fd[1]);
-	}
-	if (has_pipe(node))
-	{
-		last_pipe_fd[0] = new_pipe_fd[0];
-		last_pipe_fd[1] = new_pipe_fd[1];
 	}
 }
