+++ srcs/pipe/child.c
@@ -1,43 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils3.c                                           :+:      :+:    :+:   */
+/*   child.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/10/16 12:25:40 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/18 20:04:30 by jnakahod         ###   ########.fr       */
+/*   Created: 2021/10/18 20:02:59 by jnakahod          #+#    #+#             */
+/*   Updated: 2021/10/18 20:03:21 by jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <mini_shell.h>
 
-t_bool is_match_str(char *input, char *delimiter)
+void	child_operate_pipe_fd(t_pipe_list *first, t_pipe_list *node, int last_pipe_fd[2], int new_pipe_fd[2])
 {
-	if (!ft_strncmp(input, delimiter, ft_strlen(delimiter) + 1))
-		return (TRUE);
-	return (FALSE);
-}
-
-t_bool	is_dollar(char c)
-{
-	if (c == DLL)
-		return (TRUE);
+	if (!has_pipe(node))
+		read_pipe(last_pipe_fd);
+	else if (first == node)
+		write_pipe(new_pipe_fd);
 	else
-		return (FALSE);
-}
-
-t_bool has_heredoc(t_heredoc_to_fd *heredoc)
-{
-	if (heredoc)
-		return (TRUE);
-	else
-		return (FALSE);
-}
-
-t_bool	has_pipe(t_pipe_list *pipe_list)
-{
-	if (pipe_list->next)
-		return (TRUE);
-	return (FALSE);
+	{
+		read_pipe(last_pipe_fd);
+		write_pipe(new_pipe_fd);
+	}
 }
