+++ srcs/pipe/parent.c
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parent.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/18 20:02:01 by jnakahod          #+#    #+#             */
+/*   Updated: 2021/10/18 20:02:33 by jnakahod         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <mini_shell.h>
+
+void	parent_operate_pipe_fd(t_pipe_list *node, int last_pipe_fd[2], int new_pipe_fd[2])
+{
+	if (is_pipe_open(last_pipe_fd))
+	{
+		close(last_pipe_fd[0]);
+		close(last_pipe_fd[1]);
+	}
+	if (has_pipe(node))
+	{
+		last_pipe_fd[0] = new_pipe_fd[0];
+		last_pipe_fd[1] = new_pipe_fd[1];
+	}
+}
